add file service interface
